+++ app/src/main/AndroidManifest.xml
@@ -23,6 +23,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".presentation.test.TestActivity" />
     </application>
 
 </manifest>
